Centrally manage application syncPolicy
Remove syncPolicy setting from individual applications and manage it at the
cluster level. For an upcoming change I want the ability to globally set
`prune` to `false`, and this both enables that as well as ensuring a
consistent syncPolicy across our applications.

With this change, applications that have the `nerc.mghpcc.org/sync-policy`
label set to `common` will have their syncPolicy configured via a patch in
the cluster kustomization.yaml.