Ignore topologyupdater in nodefeaturediscovery manifests
There appears to be some weirdness around the CRD for the
NodeFeatureDiscovery operator: you need to set the `topologyUpdater`
attribute, but at least on nerc-ocp-test, when you fetch the resource, you
see `toplogyupdater` (note capitalization difference). This causes argocd
to think the app is out of sync.

This commit instructs argocd to ignore either of these names in the
resource.