Adding support for IBM autopilot GPU metrics, alerts, dashboards
We wish to use the IBM Autopilot application. A Kubernetes-native daemon
that continuously monitors and evaluates GPUs, network and storage
health, designed to detect and report infrastructure-level issues during
the lifetime of AI workloads. We will deploy autopilot to the test and
prod cluster, and the grafana dashboard to the obs cluster.